# src/main/java/ru/mentee/tasks/api/generated/controller/CommentsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech Do not edit the class manually.
 */
package ru.mentee.tasks.api.generated.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.constraints.*;
import java.util.Optional;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

@Generated(
    value = "org.openapitools.codegen.languages.SpringCodegen",
    date = "2026-01-14T22:58:36.605556+03:00[Europe/Moscow]",
    comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Comments", description = "the Comments API")
public interface CommentsApi {

  default Optional<NativeWebRequest> getRequest() {
    return Optional.empty();
  }

  /**
   * GET /tasks/{taskId}/comments : Получить комментарии к задаче
   *
   * @param taskId (required)
   * @return Список комментариев (status code 200)
   */
  @Operation(
      operationId = "getTaskComments",
      summary = "Получить комментарии к задаче",
      tags = {"Comments"},
      responses = {@ApiResponse(responseCode = "200", description = "Список комментариев")})
  @RequestMapping(method = RequestMethod.GET, value = "/tasks/{taskId}/comments")
  default ResponseEntity<Void> _getTaskComments(
      @Parameter(name = "taskId", description = "", required = true, in = ParameterIn.PATH)
          @PathVariable("taskId")
          UUID taskId) {
    return getTaskComments(taskId);
  }

  // Override this method
  default ResponseEntity<Void> getTaskComments(UUID taskId) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }
}